requirements.txt
Flask==3.0.3
PyMuPDF==1.24.0
Pillow==10.3.0
werkzeug==3.0.3
python-magic==0.4.27  # Linux file detection
rarfile==4.2
py7zr==0.21.1
pydub==0.25.1
python-docx==1.1.2
openpyxl==3.1.3
moviepy==1.0.3
gunicorn==23.0.0  # Production server